Add files via upload
This is a character that my daughter drew for me when she was 7.  My wife had a stuffed 'animal' made from the drawing and my daughter gave it to me to help keep me safe.  I worked patrol for 14 years and kept this in the car with me every day.